@@ -9,9 +9,7 @@ import 'package:moengage_flutter/geo_location.dart';
 class MoEiOSCore {
   MethodChannel _channel;
 
-  MoEiOSCore(MethodChannel channel) {
-    _channel = channel;
-  }
+  MoEiOSCore(this._channel);
 
   void enableSDKLogs() {
     _channel.invokeMethod(methodEnableSDKLogs);
