@@ -12,9 +12,6 @@ import 'package:moengage_flutter/moe_push_service.dart';
 
 Map<String, dynamic> getEventPayload(
     String eventName, MoEProperties eventAttributes) {
-  if (eventAttributes == null) {
-    eventAttributes = MoEProperties();
-  }
   Map<String, dynamic> eventPayload = eventAttributes.getEventAttributeJson();
   eventPayload[keyEventName] = eventName;
   return eventPayload;
@@ -50,7 +47,7 @@ Map<String, dynamic> getMap(String key, dynamic value) {
   return <String, dynamic>{key: value};
 }
 
-InAppCampaign inAppCampaignFromJson(dynamic methodCallArgs) {
+InAppCampaign? inAppCampaignFromJson(dynamic methodCallArgs) {
   try {
     Map<String, dynamic> inAppCampaignMap = json.decode(methodCallArgs);
 
@@ -92,7 +89,7 @@ InAppCampaign inAppCampaignFromJson(dynamic methodCallArgs) {
   return null;
 }
 
-SelfHandled selfHandledFromCampaignMap(Map<String, dynamic> inAppCampaignMap) {
+SelfHandled? selfHandledFromCampaignMap(Map<String, dynamic> inAppCampaignMap) {
   if (inAppCampaignMap.containsKey(keySelfHandled)) {
     Map<String, dynamic> selfHandledMap = inAppCampaignMap[keySelfHandled];
 
@@ -109,7 +106,7 @@ SelfHandled selfHandledFromCampaignMap(Map<String, dynamic> inAppCampaignMap) {
   }
 }
 
-NavigationAction navigationActionFromCampaignMap(
+NavigationAction? navigationActionFromCampaignMap(
     Map<String, dynamic> inAppCampaignMap) {
   if (inAppCampaignMap.containsKey(keyNavigation)) {
     Map<String, dynamic> navigationMap = inAppCampaignMap[keyNavigation];
@@ -122,7 +119,7 @@ NavigationAction navigationActionFromCampaignMap(
   }
 }
 
-CustomAction customActionFromCampaignMap(
+CustomAction? customActionFromCampaignMap(
     Map<String, dynamic> inAppCampaignMap) {
   if (inAppCampaignMap.containsKey(keyCustomAction)) {
     Map<String, dynamic> customActionMap = inAppCampaignMap[keyCustomAction];
@@ -134,7 +131,7 @@ CustomAction customActionFromCampaignMap(
   }
 }
 
-PushCampaign pushCampaignFromJson(dynamic methodCallArgs) {
+PushCampaign? pushCampaignFromJson(dynamic methodCallArgs) {
   try {
     Map<String, dynamic> pushCampaignMap = json.decode(methodCallArgs);
     String platform = pushCampaignMap.containsKey(keyPlatform)
@@ -166,10 +163,15 @@ PushCampaign pushCampaignFromJson(dynamic methodCallArgs) {
   return null;
 }
 
-PushToken pushTokenFromJson(dynamic methodCallArgs) {
-  Map<String, dynamic> pushTokenPayload = json.decode(methodCallArgs);
-  return PushToken(
-      pushTokenPayload[keyPlatform],
-      pushTokenPayload[keyPushToken],
-      MoEPushServiceExt.fromString(pushTokenPayload[keyPushService]));
+PushToken? pushTokenFromJson(dynamic methodCallArgs) {
+  try {
+    Map<String, dynamic> pushTokenPayload = json.decode(methodCallArgs);
+    return PushToken(
+        pushTokenPayload[keyPlatform],
+        pushTokenPayload[keyPushToken],
+        MoEPushServiceExt.fromString(pushTokenPayload[keyPushService]));
+  } catch (exception) {
+    print("Error: pushTokenFromJson() : ${exception.toString()}");
+  }
+  return null;
 }
